tast-tests: Add fixture to crash.Sender.mock_consent
The crash.Sender.mock_consent test currently is failing due to the
apps.SystemWebAppsInstallGuest* tests starting a Chrome instance. When
it tries to send a crash report it won't happen because the running
Chrome instance is in guest mode, this adds a precondition to ensure the
running Chrome instance is not in guest mode.

BUG=chromium:1190865
TEST=tast <amd64-DUT> apps.SystemWebAppsInstallGuest*
crash.Sender.mock_consent

Change-Id: If59a66278ca64691f51f9f53685e7b856db5c552
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2783840
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Tested-by: Ben Reich <benreich@chromium.org>
Auto-Submit: Ben Reich <benreich@chromium.org>