Add uiauto/touch
Creating touch screen based on UI has a bit of boilerplates and
it needs to keep its own internal data, so this CL creates a new
package to supply touch feature on top of uiauto.

This also consolidates the boilerplate of initializing touch
screen with coords converter. Modified ash/shelf.go and
hotseat_scroll_perf accordingly.

BUG=none
TEST=fast_build.sh -C

Change-Id: I74e6a7a0e8809254fe7d3b7c57bf84dc49d284c6