policy: fix UserAvatarImage test
Before this CL, the test is failing on some of the boards[1]. The test
is failing because red color is different from device to device.

This CL increases the maximum difference threshod from 10 to 60. This
difference represents the tolerance level to count red pixels in the
color components (rgba).

[1] https://cros-health-prod.googleplex.com/test/policy.UserAvatarImage

BUG=chromium:1185195, chromium:1191939
TEST=tast run <ip> policy.UserAvatarImage

Change-Id: Ifdfbb3d9fd224d98676343553d542eab2da4efbe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2784926
Reviewed-by: Mohamed Omar <mohamedaomar@google.com>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Commit-Queue: Mohamed Omar <mohamedaomar@google.com>
Tested-by: Mohamed Omar <mohamedaomar@google.com>