dbusutil: Move shill specific logics back to shill package.
We currently have some shill specific logic in dbusutil, but that
interface is in fact not commonly used. dbusutil might be simpler to
exclude these logics and have a shill.PropertyHolder overriding these
functions.

Also this will "fix" that the original SetProperty call of
dbusutil.PropertyHolder is implicitly shill-specific.

Pros:
- Simpler interface for both PropertyHolder.
- Reduce duplicated CreateWatcher functions in shill package.
- Allow more shill-specific utility methods.

Cons:
- Function override without an interface object might be
  confusing. (But we don't have any common utilities now, so
  there's little gain from a common {Get,Set}Property interface.)

BUG=None
TEST=go vet

Change-Id: I4f2e0f352d5983824eedfacd5a3db396436ac8ac