dbusutil: Move shill specific logics back to shill package.
We currently have some shill specific logic in dbusutil, but that
interface is in fact not commonly shared. dbusutil might be simpler to
exclude these logics and only keep the shared ones. On the other hand,
we'll have a shill.PropertyHolder overriding shill-specific functions
so the objects in shill package can still share the codes.

Also this will "fix" that the original SetProperty call of
dbusutil.PropertyHolder is implicitly shill-specific.

BUG=b:161486698
TEST=go vet
TEST=tast run $DUT wifi.SimpleConnect

Change-Id: I4f2e0f352d5983824eedfacd5a3db396436ac8ac