Extract action into its own package
I've noticed that input package depends on uiauto, which looks
strange to me. It then turend out for the Action type definition.

I think that the abstraction of Action type and its utilities
could be valuable on non-UI parts, and to avoid the unnecessary
dependencies on uiauto, the action type and its utilities should
be served from its own package.

BUG=none
TEST=fast_build.sh -C

Change-Id: I62bab8d49f895cfc82d1663cbe2932b6874d1a4e