video/platform_encoding: count number of generated key frames
Video encoders can and are configured in Chrome to generate as
few keyframes as possible (ideally only on demand) by WebRTC
(and by video.PlatformEncoding). Some hardware encoders do not
respect this command, and it'd be good to know. This CL adds
logic to platform_encoding.go to track and report that amount.

The simplest approach I found is to apply a regexp to the
hexdump output.

BUG=b:178542866
TEST=tast run <kukui|coachz|deded> video.PlatformEncoding.*_180_*
     reports correct numbers and the same as e.g.
     ffprobe  -show_frames <encoded_file> | grep key_frame=1
     (note that there's no ffmpeg on ChromeOS boards, too large).

Change-Id: I225734ba38117018c9cffe8103057b0ff5b1fb6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2785715
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Steve Cho <stevecho@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>