croshealthd/status: change timeout to be the context deadline
This change modifies the timeout for the call get cros_healthd's status
to be the context's deadline if it is set. This allows longer timeouts
if necessary.

This fixes a timeout issue for some jacuzzi devices when starting the
crosHealthdRunning fixture.

BUG=b:183228685
TEST=tast run jacuzzi-dut ui.GuestModeExtension health.ProbeNetworkInfo
(pass)
This combination of tests would previously timeout starting the
crosHealthdRunning fixture

Change-Id: If8355e20c857a9ee81cd4e9642cd728bb4bc2312
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2785716
Tested-by: Trent Begin <tbegin@chromium.org>
Auto-Submit: Trent Begin <tbegin@chromium.org>
Commit-Queue: Greg Edelston <gredelston@google.com>
Reviewed-by: Greg Edelston <gredelston@google.com>