Lacros: Promote lacros.AudioPlay
This test looks stable enough, thus promote it to be critical to prevent
regressions. The only errors are on Zork boards, but it's not related
Lacros. b/187659256 should be the issue, and culprit CL was reverted.

This only promotes this test for boards with the lacros_stable
variation. Currently, this is all non-VM boards that support lacros.

BUG=chromium:1174896
TEST=./fast_build.sh -T

Change-Id: I09958edf527744a56a33a46b725b96ef9c0c7bee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2788850
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>
Tested-by: Yuke Liao <liaoyuke@chromium.org>