Lacros: Promote lacros.AppLauncherLaunch
This test looks stable enough, thus promote it to be critical to prevent
regressions. The only errors are on Zork boards, but it's not related
Lacros. b/187659256 should be the issue, and culprit CL was reverted.

This only promotes this test for boards with the lacros_stable
variation. Currently, this is all non-VM boards that support lacros.

BUG=chromium:1174896
TEST=./fast_build.sh -T

Change-Id: I923fbf636a1abee64778cb85605438cd62719f41
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2788851
Reviewed-by: Hyung Tae Kim <hyungtaekim@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Yuke Liao <liaoyuke@chromium.org>
Commit-Queue: Yuke Liao <liaoyuke@chromium.org>