network: VPNConnect: Wait for valid SelectedService
The VPNConnect test retrieves the SelectedService property to get the
object path of the pseudo ethernet service, and then configures the
static IP of this service using the object path.

Since the pseudo device is just created at the time, the corresponding
service state could be unstable, and thus the SelectedService property
could be "/" is the service is not active, so the test code may get an
invalid service path if that happens and then the following static IP
configuration fails.

This patch changes this logic to waiting for a valid object path on
SelectedService, to avoid such cases happening.

BUG=b:165459352
TEST=The test passed 5/5 on octopus (need to run `conntrack -F` between
  two attempts, see b/182758930)

Change-Id: If0f3ab968d98408bf84df59678f47746a879ca9b