graphics.IGT: Report a PASS on skipping all tests.
Do not report an error and [FAIL] if all subtests are skipped.

Subtests could be skipped if the tests requirements are not met (i.e.
running on Zork a test that requires being on i915). Exit Status for
this case is 77.

Tast should still report an error if no tasts were run or skipped.

BUG=b:181033521
TEST=graphics.IGT.kms_big_fb for [PASS] on skip and
graphics.IGT.kms_addfb_basic for [FAIL] on real failure.

Change-Id: I9fab728b360eb25939a22b9235976e6d89f4c969