tast-tests: Simplify graphics.FastInk
The code that moves the mouse to the center of the window was needed in
Patchset 5 of the original Fast Ink Tast test CL [1] because there was a
call to ash.HideVisibleNotifications which made the mouse cursor visible
at a problematic position. Notice a code review comment there, saying I
can nix the call to ash.HideVisibleNotifications. The test does not call
ash.HideVisibleNotifications now. I can remove the code that moves the
mouse to the center of the window, and the cursor will never be shown.

[1] https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2565383/5/src/chromiumos/tast/local/bundles/cros/graphics/fast_ink.go

BUG=chromium:1070307
TEST=graphics.FastInk

Change-Id: I0d89b3ace45b269a26749ff74fde9393707f5c41