webrtc: set degradationPreference to maintain-resolution
crbug.com/1179020 fixed a bug where the tx resolution didn't adapt
given a constrained bandwidth. During tests, the default bandwidth
target is a bit small, so RTCPeerConnection starts with a small
resolution (then ramping up maybe) that falls under the sw/hw
resolution threshold, making it use the SW encoder and failing the
test.

This CL tentatively fixes that by instructing the RTCPeerConnection
to try to sustain resolution rather than quality or framerate.

'tentatively' is because I haven't been able to repro with my
local tests on hatch (kohaku) and dedede (drawlat). Discussions
in the bug seem to make this a reasonable fix though.

To do that we just set the degradationPreference [1] in the
RTCRtp(Send)Parameters dictionary used when adding the
RTCRtpTransceiver to the RTCPeerConnection.

[1] https://w3c.github.io/mst-content-hint/#degradation-preference-when-encoding

BUG=b:183408657
BUG=b:181320567
TEST=None, see above for why.

Change-Id: I8f931410c5dd6b5fc6888183a56ded6c6db94dcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2795402
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Sergey Silkin <ssilkin@chromium.org>
Reviewed-by: Alexander Hartl <alexanderhartl@google.com>
Commit-Queue: Miguel Casas <mcasas@chromium.org>