tast-tests: Pass graphics.HTMLVideoRoundedCornersUnderlay if no overlay
BEFORE: The test checks for the underlay overlay strategy, but you never
know when overlay promotion will be rejected for some irrelevant reason.
AFTER: The test checks for the underlay overlay strategy or no overlay.
The point is that the SingleOnTop overlay strategy should not be used.

BUG=chromium:1107460
TEST=graphics.HTMLVideoRoundedCornersUnderlay

Change-Id: Icf7269098b3bc0c6eace49f447d0e3b43fe92acd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2795896
Tested-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Avery Musbach <amusbach@chromium.org>