shill: add fixture to reset shill state after each test
There is some repeated code in the network tests that ensure that any
modifications to shill are reset between tests. This fixture ensures
that shill is in a known default state on start and resets it after the
test is completed. This change also adds the fixture to the
network.ConfigureServiceForProfileTest.

BUG=chromium:1178896
TEST=tast run sarien-dut network.ConfigureServiceForProfileTest (pass)

Change-Id: Ia0b241531e4f59a389b5d4c81a5161d7fbae2dc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2797632
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Yen-lin Lai <yenlinlai@chromium.org>
Reviewed-by: David Stevens <stevensd@chromium.org>
Tested-by: Trent Begin <tbegin@chromium.org>
Commit-Queue: Trent Begin <tbegin@chromium.org>