healthd/fixture.go: extend pre and post test timeouts
The PreTest and PostTest functions in the cros_healthd fixture.go check
to see that the cros_healthd service is still running after the test
runs. Some of the diagnostic routines that cros_healthd runs are very
CPU intensive (cpu_stress & cpu_cache). This can occasionally cause
these fixture functions to timeout on lower end devices. This is a
speculative fix to see if the increased timeouts are enough to get the
status after the CPU activity settles.

BUG=b:183227926
TEST=tast run sarien health.* (pass)

Change-Id: Ifd58651f545e15a4e9f6b774b3b140acfe30a134