nearbyshare: Disable SplitSettingsSync and remove unused flags
The SplitSettingsSync experiment may be affecting our in-contacts tests,
since tast's Chrome login bypasses OOBE which may result in some profile
data not being populated when the experiment is enabled. We will
explicitly disable it and see if any improvement is noted. Also cleaned
up the Nearby feature flags that are on by default on ToT.

BUG=chromium:1194707
BUG=chromium:1185910
TEST=tast run <dut> nearbyshare.*

Change-Id: I101d2e6a36f94bf9065f0ced34793eb22fbe7c38