Cache servo type to avoid re-fetching
Servo type isn't going to change mid-test. So if a test calls
servo.GetServoVersion() multiple times, it's a wasted transaction, which
costs time and potential for error.
With this CL, we cache the result for later usage.

BUG=b:177940244
TEST=Manual: Add a logging line when actually fetching the servo version
(not returning the cached value). Append to example.ServoEcho two calls
to servo.GetServoVersion. Check logs, there should only be one fetch.

Change-Id: I488f2408c3f5366dc56c2f96f50482e43e1944d8