webrtc: set start bitrate for RTCPeerConnection tests
webrtc.RTCPeerConnection tests don't set any bitrate, so by
default a conservative estimate of ~280K is used. This is OK
for QVGA streams, but we use 720p in tests and this causes
the resolution to be reduced, eventually falling back to SW
acceleration and breaking the test.

This CL adds a reasonable start bitrate, following the example
of appr.tc [1] and b/183408657#comment25

[1] https://github.com/webrtc/apprtc/blob/78600dbe205774c115cf481a091387d928c99d6a/src/web_app/js/sdputils.js#L142

BUG=b:183408657
BUG=b:181320567
TEST=webrtc.RTCPeerConnection* on samus

Change-Id: Idce19f4c4eba7933c53351b837ec9f27ad3226f1