tast: Relax CSAReconnect test conditions.
Relax the test to pass the case that the DUT eventually reconnects to
the AP even if the DUT once disconnects after receiving CSA frame. The
reason the DUT disconnects is that the Gale AP takes too long to send
beacons in the new channel so the DUT thinks the connection is lost.

BUG=b:181365077
TEST=ran tast run $DUT wifi.CSAReconnect

Change-Id: I3670a00cdf25f34b45979520f835bdd509013c20