tast: Relax CSAReconnect test conditions.
Relax the test to pass the case that the DUT eventually reconnects to
the AP even if the DUT once disconnects after receiving CSA frame. The
reason the DUT disconnects is that the Gale AP takes too long to send
beacons in the new channel so the DUT thinks the connection is lost.

The test is failing on R91, submitting this change to R91 to solve this issue.

BUG=b:181365077
TEST=ran tast run $DUT wifi.CSAReconnect

Change-Id: I3670a00cdf25f34b45979520f835bdd509013c20
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2802698
Tested-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>
(cherry picked from commit 68047f33b024fc1d19148dc602b529f4eb42d972)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2853424