webrtc: fix copy-paste error in bitrate setting JS method
crrev/c/2802691 landed code to set the initial bitrate for
the WebRTC RTCPeerConnections used in tests and by (my)
mistake, I copy-pasted the method call, however a parameter
is undefined in one of the callsites and the test fails.

This CL corrects that by using 'VP8' as intended (simulcast
is only defined ATM for VP8 in these tests).

BUG=b:183408657
BUG=b:181320567
TEST=webrtc.RTCPeerConnection.vp8_simulcast on trogdor

Change-Id: Ie6b03073e2d97fea3b9fdc542bfab2c4e00deaf7