graphics/perf: add Package C-state residency collection
This CL adds collection of the % of time the SoC spent in one of
several sleep modes (called C-States) during perf tests. Package
C-State residencies are reported via "perf" as the amount of
cycles the package was in them during the collection time -- with
the total amount of cycles to be found in yet another perf event
called "TSC" for Time Stamp Counter.

The code is added to graphics/perf.go where other perf collection
happens. I tried to refactor a bit more the two perf calls but
since the GPU ones are retrieved as time and the C-states are
reported as numbers and for the first CPU of the package, it
turned out that refactoring made the code quite confusing. Hence
the apparent code duplication, which is not that egregious upon
further closer inspection.

ATM only Intel platforms support these PMU counters, but the
kernel interface is generic (or it should be).

BUG=b:184842379
TEST=tast run <volteer-IP> video.PlaybackPerf.h264_2160p_60fps_hw:
     [17:17:35.370] Measuring GPU usage for 25s
***  [17:17:35.371] Measuring Package C-State residency for 25s
     [17:17:40.371] Measuring CPU usage for 25s
     [17:18:00.396] Average frequency: 197.227790MHz
     [17:18:00.396] rcs: 4.339726%
     [17:18:00.396] vcs: 18.507167%
     [17:18:00.396] rc6: 77.190391%
***  [17:18:00.396] c2: 14.046883%
***  [17:18:00.397] c3: 0.146370%
***  [17:18:00.397] c6: 0.455601%
***  [17:18:00.397] c7: 16.718243%
***  [17:18:00.397] c8: 4.915800%
***  [17:18:00.397] c9: 0.000000%
***  [17:18:00.397] c10: 0.000000%
***  [17:18:00.397] c0: 63.717103%
     [17:18:05.379] CPU usage: 5.652527%
     [17:18:05.379] Avg pkg power usage: 2.128330W

(New metrics marked with ***)
Also verified on trogdor: no metrics but no error.

Change-Id: I84d5fe1ded8030d9b3c975c220fb142829ef3bcc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2819190
Tested-by: Miguel Casas <mcasas@chromium.org>
Reviewed-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Reviewed-by: Ilja H. Friedel <ihf@chromium.org>
Commit-Queue: Miguel Casas <mcasas@chromium.org>