chrome: Ignore errors on releasing `this` object
Conn.Call captures the `this` object, call a function on it, then
releases the reference to the `this` object. However, if an evaluated
expression causes navigation or browser restart, release may fail.

This patch updates the method to ignore errors on release.

BUG=chromium:1193417
BUG=chromium:1197692
TEST=tast run $DUT ui.ChromeLogin

Change-Id: I1cf3d526d65875aca01848cf3cd5eff0a16b4f77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2820576
Tested-by: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Mitsuru Oshima (Slow - Gardening) <oshima@chromium.org>