tast: Use new property to check android boot status
Replace the usage of net.tcp.default_init_rwnd with ro.arc.on_boot when
android on arcvm. When tast tests are run on arcvm, ro.arc.on_boot is
set when android init reaches "on boot" event. Function WaitAndroidInit
check the appropriate property to determine if android boot
successfully. If the function does not get the property, it terminates
the test, otherwise, test continues.

BUG=b:184004893
TEST=Ran arc.Boot.vm and arc.Boot

Change-Id: I1f2f13bc5eeed79280c925b1602977dac1177b2f