nearbyshare: temporarily disable verbose chrome logging
The issue we are seeing with /var/log/chrome/chrome becoming a regular
file started happening in the same build where we enabled verbose chrome
logging. I want to try removing these flags temporarily to see if
they're causing the issue. We can reenable them if they turn out not to
be, or otherwise figure out a way to enable them without breaking the
tests.

BUG=chromium:1197197
TEST=tast run <dut> nearbyshare.*
Exempt-From-Owner-Approval: just removing some chrome args, want to verify this quickly in the lab

Change-Id: Ifcfdc71dce521f6aaf32ce07ed4d4a1408113c9f