camera: Ensure the window of app is closed while closing app
There is a flakiness when we run stress tests for camera app. The reason
is that sometimes when restarting app, it will use the previous browser
to load the camera app instead of creating a new one, which is
unexpected. It might happen because the actual destruction of the window
is asynchronous.

By polling for autotestPrivate.isSystemWebAppOpened() API, we can ensure
that the window of the app is closed while closing app.

The corresponding Chrome CL is: https://crrev.com/c/2820446

BUG=b:184131032
TEST=tast run [DUT] camera.CCAUIStress.fake

Change-Id: I95cbda53e06eeaed0ff298c3ab00a618bb09a26f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2825771
Tested-by: Wei Lee <wtlee@chromium.org>
Commit-Queue: Wei Lee <wtlee@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Reviewed-by: Shik Chen <shik@chromium.org>