platform.BootPerf: Measure time from OS to kernel shutdown
By looking at the timestamp of the last line of ramoops, we can
get a good estimate of the time when the kernel requests the FW
to reboot.

This allows us to compute the time between OS requesting reboot,
and the kernel requesting reboot from the firmware.

Note that we assume that kmsg timestamp is equivalent to
CLOCK_BOOTTIME, which is not 100% correct (see bug for data).
We may want to refine the computation method (by synchronizing
the clocks), if this is something we want to measure precisely.

BUG=b:181084548
TEST=tast run -var=platform.BootPerf.iterations=10 \
              -var=platform.BootPerf.skipRootfsCheck=true \
              $IP platform.BootPerf

Change-Id: I0356ef43cb6d6dbed3ed892d22a36a6093c078cc