Add arc.DefaultDisplayDensity tast test
Add a tast test to ensure the display density (scale) is calculated and
set correctly on boot for the internal display on various boards.

Android's density is calculated from Chrome's device_scale_factor, with
special considerations for low-density devices. The density being used
must be one of the approved screen densities listed in the Android CDD.

This test calculates the expected density value for the device, and
verifies that it is set correctly by checking the override density
output in dumpsys.

BUG=b:184362786
TEST=tast run DUT arc.DefaultDisplayDensity

Change-Id: I495f9cf959897d9e419b20f0a40e882bb07340d8