camera: Use separate timeout for subtests in CCAUI*
Originally, when a subtest hangs and reaches the timeout of the test,
all following subtests will fail with timeout error and make it
difficult to understand the root cause of the test failures.

We should separate the timeouts for each sub tests.

Besides that, this CL also refactor camera.CCAUISettings and
camera.CCAUITakePicture a bit using subtests.

BUG=b:184131041
TEST=tast run [DUT] camera.CCAUI* (on winky and banjo)

Change-Id: I4e0b40e984bc35c3f6bc036ba405aa54b322a7fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2825773
Tested-by: Wei Lee <wtlee@chromium.org>
Commit-Queue: Wei Lee <wtlee@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>