servo: Check whether control is supported
In Autotest, we check whether a servo control is supported via the `doc`
method. If `doc controlName` returns a fault, then controlName is not
supported.

This CL adds the same logic to Tast. This includes adding handling for
XML-RPC faults returned by Servo.

BUG=b:177940244
TEST=fast_build.sh
TEST=Manual: Call HasControl on a known-good control like "ec_uart_cmd",
and on a known-bogus control like "foo".

Change-Id: I89e2c760b91cc21a20f504221c75d128e7503531