network.DiagFailFirewall: add test
This test creates a firewall that blocks traffic on specified ports and
runs the network diagnostic firewall routines to ensure the firewall is
detected. It also modifies the CreateFirewall function to add a timeout
that is passed to the iptables command. This allows the command to wait
until other operations are complete.

BUG=chromium:1178896
TEST=tast run hatch-dut network.DiagFailFirewall* (pass)

Change-Id: I6755aa4341f7ce2b487fff6572d9f9bac8beb77f