tast-tests: convert PS175 firmware install to detect
Changes to how fwupd uses flashrom mean it is no longer possible to mock
flashrom as the test was doing, and the PS175 has different IDs. Change
the test so it ensures fwupd detects the device, but does not attempt to
update it.

BUG=b:173742142
TEST=Test now passes on Noibat

Cq-Depend: chromium:2829173
Change-Id: I61d31790c498c11f7d34ce9994427cfea17d139e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2825777
Commit-Queue: Daniel Campello <campello@chromium.org>
Reviewed-by: Daniel Campello <campello@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Daniel Campello <campello@chromium.org>