Add hardware dependencies to ARC input tast tests
Some input tests for ARC++ have hardware requirements like a touchscreen
an internal display.

arc.TouchInput should only run on devices with a built-in touchscreen.

arc.MouseInput is broken into two tests: arc.MouseInput.clamshell and
arc.MouseInput.tablet, where the latter requires the presence of an
internal display so that it does not run on Chromeboxes/Chromebits.

arc.MouseInput is also now disabled for P, and will only run on VM. This
is because the test for P does not provide any valueable information and
is flaky, and since this is a heavy and long test, it will reduce the
load on the labs.

BUG=b:162923702
TEST=tast run DUT arc.{MouseInput.tablet, MouseInput.clamshell,
TouchInput}

Change-Id: Id04336271fd8111d454ec13ecf655a37c4065cb0