video/video.html: Make window contain the video.
This CL makes video.html render the video similarly to how it would be
rendered if the video had been opened directly in Chrome. For example,
if the video is large, it scales it down so that the window contains the
full video (instead of cropping the video).

I simply opened a video in Chrome on my Chromebook and inspected the
html, body, and video elements. Then, I just copied and pasted the CSS
of each element into video.html. I also copied and pasted a <meta>
element. Then, I deleted things to minimize the CSS while still getting
the video on an overlay.

Doing this ensures that if the video got promoted to an overlay before
CL:2770281, it also gets promoted after this CL.

BUG=b:185164621,b:185165312
TEST=video.PlaybackPerf.h264_1080p_60fps_sw on affected board

Change-Id: Ie1dc87ad48b2fd0d4eb75903eb82bdc4d7c35d7b