Fix arc.PointerCapture by waiting for window focus
There are two issues addressed by this CL:

- UIAutomator cannot access Android apps when a Chrome window is in
focus because in Android, the PlaceholderActivity takes up the full
screen. To get around this, we instead check that the test activity
loses focus when a Chrome window is focused. This CL adds a way to query
whether an Activity is focused in a display.

- There is a bug (b/185837950) where the first mouse event is consumed
by Chrome after enabling Pointer Capture with a keyboard. To get around
this, we send an extra mouse event in the affected tests.

BUG=b:185837950
TEST=tast run DUT arc.PointerCapture

Change-Id: I694bc77a38d967041146f6f2e8c772c62e1dba4e