Fix multivm memory metrics < 4.19
Older kernels have fewer fields in zram mm_stats, so make the newer
fields optional. Upstream will introduce a 9th field, so add that in,
and support (but don't parse) any extras so we don't break in the future
if more are added.

Grunt boards have processes that contain empty smaps_rollup files, so
ignore these instead of erroring out. I am not sure if this has
something to do with the kernel version, but it should be fixed to
support grunt.

BUG=b:185430646
TEST=tast run { hatch, grunt } multivm.Login.novm

Change-Id: I03133fc3f000525d1ff0a44ad7d2bd36280ab77a