tast-tests: Add arc.MediaProjectionPermissions
When using the MediaProjection API in ARC++/ARCVM, we use the Desktop
Media Picker in Chrome, instead of the Android permissions dialog, to
handle permissions. A notification in the system tray is also shown
during screen capture. This tast test ensures that we are using the
Chrome dialog and showing the notification. It does not cover screen
capture testing because that is covered by CTS.

BUG=b:155006957
TEST=tast run <DUT> arc.MediaProjectionPermissions

Change-Id: I2662a98ea12d22d295aeb86fd2bb478b57b81c2a