nearbyshare: enable verbose logs for bluetooth and nearby modules
Enabling --v=1 had some unforeseen issues (crbug/1197197). Let's try a
more limited set of verbose logs and see if that will hold us over for
now. The log sizes after this change are very close to the current
sizes, so this shouldn't explode the chrome logs and make us hit that
issue again.

BUG=chromium:1200948
TEST=nearbyshare.*

Change-Id: I6cf940e8c6392f208da2d8c441a034691ad15ff1