Tast: Add Tast test for in-session EDU coexistence
This CL adds a Tast test for the in-session EDU coexistence feature.
The test logs in as a Unicorn account and adds an EDU secondary
account, which is the only available secondary account type for
Unicorn accounts. The test also verifies that the K-12 EDU secondary
account was added successfully.

BUG=chromium:1204264
TEST=Tast test passes locally

Change-Id: Iaf592f79377896b7b6a71bb2d6e9917bcac83752