Make platform.MemoryPressure.vm actually run a VM
It doesn't make sense that the VM flavour of the test does not actually
run the VM, unless an addition argument is set.

This CL gets rid of the argument, and has the param control whether a VM
(or android_p container) is run during the test.

This change will cause this test to run on more boards, as the default
param will run even without android_p or android_vm SoftwareDeps.

BUG=b:186072906
TEST=tast run hatch platform.MemoryPressure*

Change-Id: Id9a4e11185f11b39dba4979a348b0dba275b803e