nearbyshare: update UI params for the incoming share notification
"Receive" was changed to "Accept" for the notification button to accept
an incoming share, which is why the Phone->CB tests started failing.
CB->CB tests were still passing since the receiver CB is usually running
an M91 build at the moment with the way things are scheduled (the DUTs
are not provisioned with the same version before the tests run due to
lack of multi-DUT support). Due to this version skew in CB->CB tests,
we'll have to check for either string instead of simply updating it to
"ACCEPT".

BUG=chromium:1201855
TEST=tast run <dut> nearbyshare.PhoneToCrosInContacts*

Change-Id: I16c85c6ab94886d67c0c028ddaa074f5a7f2cd8b