arcappcompat: Update Tast test for Netflix
BUG=b:186241114
TEST=tast run <eve> arcappcompat.Netflix

Netflix is getting failed in launch. It is observed that selecting the “OK” option for “Download and Go” dialog box doesn’t appear on all devices. Also observed that there is a new window and a pop up for Clamshell:Signout testcase. Updated the code to fix it.

Change-Id: I562836f832d79a41c19da00da3ae1c3dafa46b3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2848783
Tested-by: Meenakshi Thiyagarajan <mthiyagarajan@google.com>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Meenakshi Thiyagarajan <mthiyagarajan@google.com>