scanner: Stop lorgnette between EnumerateIPPUSB iterations
The test currently stops lorgnette at the very end of the test.  Since
it takes time to restart the virtual-usb-printer at the beginning of
each iteration, this seems to occasionally flake when d-bus tries to
start lorgnette at the exact same time that lorgnette is timing out from
the previous iteration.  We want lorgnette to be in a consistent state
for each iteration anyway, so move the StopService call into the inner
iteration instead of the outer test end.

BUG=b:186569183
TEST=15 iterations of scanner.EnumerateIPPUSB on betty-arc64

Change-Id: I829deb973bc0b5b6db641fa0f6f8e0eb0d8d033f