security.Mtab: Temporarily ignore "proc" mount types.
On kernel 3.18, nsfs mounts (bind-mounts of files in
/proc/<pid>/ns/<ns_type>) show up as type "proc" instead of type
"nsfs". This is making it hard to land some security improvements.

Ignore "proc" mounts for now, in order to land these improvements
(and remove the entry for mnt_chrome since this will be either "nsfs"
or "proc" and therefore always ignored.)

A subsequent CL will refactor this test to make it less fragile:
-Concentrate on preventing W+X mounts.
-Reduce the set of filesystems that we keep in the baseline by:
--Automatically allowing filesystems that show up with default RO or RW options.

Also, take this opportunity to remove a baseline entry corresponding to
a fixed bug.

BUG=chromium:1204115
TEST=Passes on betty-pi-arc.

Change-Id: Ia432e75939b19a4f175757ffa8bc41303da529ac