health.ProbeNetworkInfo: poll routine
This change adds polling to the network health info telemetry request.
If this routine is run right after chrome starts, it's possible that the
network health information has not yet been populated. If there is any
other error besides empty fields, the routine will immediately report an
error and stop.

BUG=chromium:1202124
TEST=tast run kevin-dut health.ProbeNetworkInfo (pass)

Change-Id: Iff080755be7c87d66e4d44c749faf96d8c438471